lsa_serialise: Use size_t for lengths, and add arithmetic overflow checks.
Reported-by: Ronald Huizer <r.huizer@xs4all.nl>

+++ lsa_serialise.h
@@ -24,11 +24,11 @@
 
 #define MAX_SERIALISED_INT_LEN		10
 
-int lsa_serialise_length(struct lsa *lsa, uint8_t *preid);
-int lsa_serialise(uint8_t *buf, int buflen, int serlen,
-		  struct lsa *lsa, uint8_t *preid);
-int lsa_attr_serialise_length(struct lsa_attr *attr);
-int lsa_attr_serialise(uint8_t *buf, int buflen, struct lsa_attr *attr);
+size_t lsa_serialise_length(struct lsa *lsa, uint8_t *preid);
+size_t lsa_serialise(uint8_t *buf, size_t buflen, size_t serlen,
+		     struct lsa *lsa, uint8_t *preid);
+size_t lsa_attr_serialise_length(struct lsa_attr *attr);
+size_t lsa_attr_serialise(uint8_t *buf, size_t buflen, struct lsa_attr *attr);
 
 
 #endif
